requirements.txt
Django==3.2.8
Pillow==8.4.0
python-dateutil==2.8.1
pytz==2020.1
psycopg2-binary
django-summernote==0.8.11.6
django-anymail==8.1
django-bootstrap5
django-bootstrap4
django-debug-toolbar==3.2.1
django-two-factor-auth
phonenumbers
git+https://github.com/dqprogramming/django-fluid-permissions@master#django-fluid-permissions
django-bleach
django-ckeditor
-e git+https://github.com/dqprogramming/thoth-client.git#egg=thothlibrary
python-crontab==2.2.3
django-basicauth==0.5.3
django-request
django-crispy-forms
tqdm
pycountry
babel
djangorestframework
python-magic==0.4.25
resize-image
python-magic
django-extensions
babel
gunicorn
django-middleware-global-request==0.2.0
boto3
rich
humanize
sentry-sdk
python-docx==0.8.11
CurrencyConverter==0.17.5
requests
pdfkit
docx